+++ packages/transforms/README.md
@@ -1,1 +1,10 @@
 # transforms
+
+## features
+
+- server reference
+  - [x] "use server" file
+  - [x] "use server" inline
+  - [ ] closure encryption
+- client reference
+  - [ ] "use client" file
