+++ packages/transforms/src/types/estree.ts
@@ -1,0 +1,9 @@
+import type {} from "estree";
+
+// rollup ast has node position
+declare module "estree" {
+  interface BaseNode {
+    start: number;
+    end: number;
+  }
+}
